Allow serializing SM state
It is useful in some applications to get a snapshot of the stream
management state for storage outside the process, in order to recover
from crashes and other things.  The get_sm_state already present is not
suitable for this because it returns a live object tied in to the
current context and such, and containing much unneeded internal-api
data.

Introduce a new get_sm_serializable_state that creates instead a fixed
snapshot, handing full memory ownership to the caller, and a dual
set_sm_serializable_state which takes in the bare minimum needed values
in order to resume and sets up a new sm_state based on these values.