Callbacks when a stanza is ack'd or fails to send
This adds two new callbacks, one of which runs with a stanza id when a
stanza sends and gets an sm ack from the server, and one which runs with
a stanza id when a stanza fails and will not be retried.

Incomplete and presented for discussion of the design.

An id element has been added to the send/sm queue in order to facilitate
passing this to the callback without re-parsing the stanza.

This is also added to the sm serialization, though we could parse the
stanza on restore to avoid this I suppose.

TODO: fail everything on stream end